Add class for GraphUtils
